@@ -1,76 +1,68 @@
 name: alheekmahlib_website
 description: A new Flutter project.
 
-
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
 
 version: 2.0.0+2
 
 environment:
-  sdk: '>=3.0.6 <4.0.0'
-
+  sdk: ^3.6.1
 
 dependencies:
+  animated_toggle_switch: ^0.8.5
+  animations: ^2.1.0
+  another_xlider: ^3.0.2
+  arabic_numbers: ^1.0.0
+  auto_size_text: ^3.0.0
+  bot_toast: ^4.1.3
+  clipboard: ^2.0.2
+  cupertino_icons: ^1.0.8
+  easy_image_viewer: ^1.5.1
+  expansion_tile_card: ^3.0.0
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+  flutter_native_splash: ^2.4.6
+  flutter_sliding_up_panel: ^2.1.1
+  flutter_staggered_animations: ^1.1.1
+  flutter_staggered_grid_view: ^0.7.0
+  flutter_svg: ^2.2.1
   flutter_web_plugins:
     sdk: flutter
-  intl: ^0.18.1
-
-
-
-  cupertino_icons: ^1.0.6
-  theme_provider: ^0.6.0
-  flutter_svg: ^2.0.8
-  flutter_staggered_animations: ^1.1.1
-  shared_preferences: ^2.2.2
-  another_xlider: ^3.0.1
-  #  dropdown_button2: ^2.1.3
-  get: ^4.6.6
-  lottie: ^2.7.0
-  clipboard: ^0.1.3
-  share_plus: ^7.2.1
-  url_launcher: ^6.2.1
-  animations: ^2.0.8
-  infinite_carousel: ^1.0.3
+  gap: ^3.0.1
+  get: ^4.7.2
+  get_it: ^8.2.0
+  go_router: ^16.2.4
   hijri: ^3.0.0
-  arabic_numbers: ^1.0.0
-  sliding_up_panel: ^2.0.0+1
-  auto_size_text: ^3.0.0
-  flutter_native_splash: ^2.3.5
-  get_it: ^7.6.4
-  flutter_slider_drawer: ^2.1.3
-  expansion_tile_card: ^3.0.0
+  http: ^1.5.0
+  infinite_carousel: ^1.1.1
+  intl: ^0.20.2
+  just_audio: ^0.10.5
+  lottie: ^3.3.2
+  rxdart: ^0.28.0
   scrollable_positioned_list: ^0.3.8
-  http: ^1.1.0
-  easy_image_viewer: ^1.2.1
-  just_audio: ^0.9.35
-  bot_toast: ^4.1.3
-  flutter_sliding_up_panel: ^2.1.1
-  animated_toggle_switch: ^0.8.0
-  rxdart: ^0.27.7
-  go_router: ^12.0.1
   seo_renderer: ^0.6.0
-
-
+  share_plus: ^12.0.0
+  shared_preferences: ^2.5.3
+  sidebarx: 0.17.2
+  sliding_up_panel: ^2.0.0+1
+  url_launcher: ^6.3.2
+  google_huawei_availability:
+    git:
+      ref: dev
+      url: https://github.com/alheekmahlib/google_huawei_availability.git
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
 
   flutter_lints: ^3.0.0
-
+  flutter_test:
+    sdk: flutter
 
 flutter:
   generate: true
 
-
   uses-material-design: true
-
 
   assets:
     - assets/svg/
@@ -85,8 +77,6 @@ flutter:
     - assets/svg/juz/
     - assets/json/translate/
   #   - images/a_dot_ham.jpeg
-
-
   fonts:
     - family: uthmanic
       fonts:
@@ -108,6 +98,15 @@ flutter:
     - family: naskh
       fonts:
         - asset: assets/fonts/NotoNaskhArabic-VariableFont_wght.ttf
+    - family: cairo
+      fonts:
+        - asset: assets/fonts/Cairo-Bold.ttf
+          weight: 900
+        - asset: assets/fonts/Cairo-SemiBold.ttf
+          weight: 600
+        - asset: assets/fonts/Cairo-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/Cairo-Regular.ttf
 
 flutter_native_splash:
   color_web: "#f3efdf"
