@@ -1,35 +1,55 @@
 import 'package:flutter/material.dart';
+import 'package:gap/gap.dart';
 
-import '/features/home_screen/widgets/about_lib.dart';
 import '/features/home_screen/widgets/our_apps.dart';
-import 'widgets/dwaa_daily.dart';
+import '../../core/services/controllers/general_controller.dart';
+import '../../services_locator.dart';
+import 'widgets/about_us_section.dart';
+import 'widgets/faq_section.dart';
+import 'widgets/hero_header.dart';
+import 'widgets/services_section.dart';
 
 class HomeScreen extends StatelessWidget {
-  const HomeScreen({Key? key}) : super(key: key);
+  const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     double height = MediaQuery.sizeOf(context).height;
     double width = MediaQuery.sizeOf(context).width;
+
     return Directionality(
       textDirection: TextDirection.rtl,
       child: Container(
         height: height,
         width: width,
-        color: Theme.of(context).colorScheme.background,
-        child: const SingleChildScrollView(
+        color: Theme.of(context).colorScheme.surface,
+        child: SingleChildScrollView(
           child: Padding(
-            padding: EdgeInsets.only(top: 64.0, bottom: 32.0),
+            padding: const EdgeInsets.only(top: 32.0, bottom: 48.0),
             child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
-                Wrap(
-                  alignment: WrapAlignment.center,
-                  children: [
-                    SizedBox(width: 350, child: AboutLib()),
-                    SizedBox(width: 350, child: DwaaDaily()),
-                  ],
+                const Padding(
+                  padding: EdgeInsets.symmetric(horizontal: 16.0),
+                  child: HeroHeader(),
                 ),
-                OurApps(),
+                const Gap(32),
+                const Padding(
+                  padding: EdgeInsets.symmetric(horizontal: 16.0),
+                  child: ServicesSection(),
+                ),
+                const Gap(32),
+                KeyedSubtree(
+                    key: sl<GeneralController>().ourAppsKey,
+                    child: const OurApps()),
+                const Gap(32),
+                const AboutUsSection(),
+                const Gap(32),
+                const Padding(
+                  padding: EdgeInsets.symmetric(horizontal: 16.0),
+                  child: FaqSection(),
+                ),
+                const Gap(32),
               ],
             ),
           ),
