@@ -1,12 +1,10 @@
 import 'package:shared_preferences/shared_preferences.dart';
 
-
 class SettingsHelpers {
-
   SharedPreferences? prefs;
 
   static final SettingsHelpers instance = SettingsHelpers._internal();
-  SettingsHelpers._internal(){
+  SettingsHelpers._internal() {
     _init();
   }
 
@@ -23,8 +21,10 @@ class SettingsHelpers {
   static const double minFontSizeArabic = 20;
 
   double? get getFontSizeArabic {
-    String fontSizeString = prefs!.getString('fontSizeArabic')!;
-    return double.tryParse(fontSizeString ?? minFontSizeArabic.toString());
+    final fontSizeString = prefs?.getString('fontSizeArabic');
+    if (fontSizeString == null || fontSizeString.isEmpty) {
+      return minFontSizeArabic;
+    }
+    return double.tryParse(fontSizeString) ?? minFontSizeArabic;
   }
 }
-
