@@ -1,36 +1,108 @@
-import 'package:alheekmahlib_website/core/utils/constants/extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+
+import '../../../core/utils/helpers/app_router.dart';
+import '../../../services_locator.dart';
 
 class BottomBar extends StatelessWidget {
   const BottomBar({super.key});
 
   @override
   Widget build(BuildContext context) {
+    final scheme = Theme.of(context).colorScheme;
+    final width = MediaQuery.sizeOf(context).width;
+    final isWide = width > 820;
+    final year = DateTime.now().year.toString();
+
     return Container(
-      height: 40,
-      width: MediaQuery.sizeOf(context).width,
+      height: 50,
+      width: width,
       decoration: BoxDecoration(
-          color: Theme.of(context).primaryColorDark,
-          borderRadius: const BorderRadius.only(
-            topLeft: Radius.circular(8.0),
-            topRight: Radius.circular(8.0),
-          ),
-          boxShadow: [
-            BoxShadow(
-                color: context.textDarkColor.withOpacity(.3),
-                blurRadius: 20,
-                offset: const Offset(0, -2))
-          ]),
-      alignment: Alignment.centerLeft,
+        color: scheme.surface,
+        borderRadius: const BorderRadius.only(
+          topLeft: Radius.circular(12.0),
+          topRight: Radius.circular(12.0),
+        ),
+        border: Border(
+          top: BorderSide(color: scheme.outlineVariant, width: 1),
+        ),
+      ),
       padding: const EdgeInsets.symmetric(horizontal: 16.0),
-      child: Text(
-        '${'appName'.tr} | ١٤٤٥ هـ',
-        style: TextStyle(
-            color: context.iconsDarkColor,
-            fontFamily: 'kufi',
-            fontSize: 11,
-            height: 1.5),
+      child: SafeArea(
+        top: false,
+        child: Center(
+          child: ConstrainedBox(
+            constraints: const BoxConstraints(maxWidth: 1100),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment:
+                  isWide ? CrossAxisAlignment.center : CrossAxisAlignment.start,
+              children: [
+                SizedBox(height: isWide ? 6 : 4),
+                // CTA + حقوق النشر
+                Row(
+                  mainAxisAlignment: isWide
+                      ? MainAxisAlignment.spaceBetween
+                      : MainAxisAlignment.start,
+                  children: [
+                    // حقوق النشر
+                    Row(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Icon(Icons.copyright,
+                            size: 14, color: scheme.onSurfaceVariant),
+                        const SizedBox(width: 6),
+                        Text(
+                          '${'appName'.tr} • $year',
+                          style: TextStyle(
+                            color: scheme.onSurfaceVariant,
+                            fontFamily: 'cairo',
+                            fontSize: 12,
+                            height: 1.3,
+                          ),
+                        ),
+                      ],
+                    ),
+                    const SizedBox(width: 12),
+                    // زر رئيسي لبدء المشروع
+                    if (isWide)
+                      ElevatedButton.icon(
+                        onPressed: () =>
+                            sl<AppRouter>().onItemTapped(4, context),
+                        icon: const Icon(Icons.send_outlined, size: 16),
+                        label: Text(
+                          'cta_start_project'.tr,
+                          style: const TextStyle(fontFamily: 'kufi'),
+                        ),
+                        style: ElevatedButton.styleFrom(
+                          padding: const EdgeInsets.symmetric(
+                              horizontal: 14, vertical: 10),
+                          shape: const StadiumBorder(),
+                        ),
+                      )
+                    else
+                      SizedBox(
+                        width: double.infinity,
+                        child: ElevatedButton(
+                          onPressed: () =>
+                              sl<AppRouter>().onItemTapped(4, context),
+                          style: ElevatedButton.styleFrom(
+                            padding: const EdgeInsets.symmetric(
+                                horizontal: 14, vertical: 12),
+                            shape: const StadiumBorder(),
+                          ),
+                          child: Text(
+                            'cta_start_project'.tr,
+                            style: const TextStyle(fontFamily: 'kufi'),
+                          ),
+                        ),
+                      ),
+                  ],
+                ),
+              ],
+            ),
+          ),
+        ),
       ),
     );
   }
