@@ -4,17 +4,19 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '/core/services/controllers/general_controller.dart';
 import 'core/services/controllers/apps_info_controller.dart';
-import 'core/services/controllers/athkar_controller.dart';
 import 'core/services/controllers/audio_controller.dart';
 import 'core/services/controllers/ayat_controller.dart';
 import 'core/services/controllers/books_controller.dart';
+import 'core/services/controllers/contact_controller.dart';
 import 'core/services/controllers/details_screen_controller.dart';
-import 'core/services/controllers/pageNavigation_controller.dart';
 import 'core/services/controllers/quranText_controller.dart';
 import 'core/services/controllers/settings_controller.dart';
 import 'core/services/controllers/surahTextController.dart';
+import 'core/services/controllers/theme_controller.dart';
 import 'core/services/controllers/translate_controller.dart';
 import 'core/services/shared_pref_services.dart';
+import 'core/utils/helpers/app_router.dart';
+import 'features/athkar_screen/controllers/athkar_controller.dart';
 
 final sl = GetIt.instance;
 
@@ -70,8 +72,17 @@ class ServicesLocator {
         Get.put<DetailsScreenController>(DetailsScreenController(),
             permanent: true));
 
-    sl.registerLazySingleton<PageNavigationController>(() =>
-        Get.put<PageNavigationController>(PageNavigationController(),
-            permanent: true));
+    sl.registerLazySingleton<AppRouter>(
+        () => Get.put<AppRouter>(AppRouter(), permanent: true));
+
+    // Eager registration to ensure availability for Get.find in widgets
+    final ContactController contactController =
+        Get.put<ContactController>(ContactController(), permanent: true);
+    sl.registerSingleton<ContactController>(contactController);
+
+    // Theme controller must be available before building MyApp (used by GetBuilder)
+    final ThemeController themeController =
+        Get.put<ThemeController>(ThemeController(), permanent: true);
+    sl.registerSingleton<ThemeController>(themeController);
   }
 }
